Update dependencies and integrate Podfile for iOS
Updated the dependencies in the pubspec.yaml and pubspec.lock files which includes changes like removing 'cupertino_icons' and adding 'crypto' as a transitive dependency. Also, a Podfile for iOS was integrated to support Cocoapods. This commit is significant for maintaining the relevancy of the app dependencies and improving compatibility with iOS.